[chore] ExecuteCommandUseCase에 커맨드 실행을 웹소켓으로 하는 메서드 추가
